trying to work on home-routes and file structure, editing routesand js files
