+++ controllers/home-routes.js
@@ -1,26 +1,37 @@
 const router = require('express').Router();
-// var session = require('express-session');
-// const { user, mood } = require('../models');
+const { User } = require('../models');
+
+// GEt ROUTES
+
 
 // SignUp/Login
-router.get('/', (req, res) => {
-    res.render('login');
+
+
+
+
+
+
+
+// router.get('/', (req, res) => {
+//     res.sendFile(path.join(__dirname, "../homepage"));
+// });
 // Homepage
-res.render('homepage', {
-        // moods? 
-            loggedIn: req.session.loggedIn,
-          })
-});
+// res.render('homepage', {
+//         // moods? 
+//             loggedIn: req.session.loggedIn,
+//           })
+// });
 
 
 // OR
 
-// router.get('/', (req, res) => {
-//     res.render('homepage', {
-//         // moods? 
-//         loggedIn: req.session.loggedIn,
-//       })
-// });
+router.get('/login.js', async (req, res) => {
+    res.render('login');
+})
+
+router.get('/login.js', (req, res) => {
+    res.render('signup');
+});
