+++ public/js/logout.js
@@ -1,16 +1,16 @@
-// Logout from the application
-const logout = async () => {
-    const response = await fetch('/api/users/logout', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-    });
+// // Logout from the application
+// const logout = async () => {
+//     const response = await fetch('/api/users/logout', {
+//       method: 'POST',
+//       headers: { 'Content-Type': 'application/json' },
+//     });
   
-    if (response.ok) {
-      document.location.replace('/');
-    } else {
-      alert('Failed to log out.');
-    }
-  };
+//     if (response.ok) {
+//       document.location.replace('/login');
+//     } else {
+//       alert('Failed to log out.');
+//     }
+//   };
   
-  document.querySelector('#logout').addEventListener('click', logout);
+//   document.querySelector('#logout').addEventListener('click', logout);
