+++ views/layouts/partials/navbar.handlebars
