+++ views/login.handlebars
@@ -84,7 +84,7 @@
         </form>
     </section>
 
-    <script src="/js/signUp.js"></script>
     <script src="/js/login.js"></script>
+    <script src="/js/logout.js"></script>
 
 </body>
